Working on implementing an observer pattern for all estimators. The observer for source events is complete but not tested.

+++ packages/facemc/core/src/Facemc_ModuleTraits.hpp
@@ -31,7 +31,7 @@ namespace Facemc{
 struct ModuleTraits
 {
   //! The internal source handle type
-  typedef unsigned long long InternalSourceHandle;
+  typedef unsigned int InternalSourceHandle;
   //! The internal estimator handle type
   typedef unsigned long long InternalEstimatorHandle;
   //! The internal material handle type
